requirements.txt
-e .
python-telegram-bot>=10.1.0
RPi.GPIO>=0.6.3
git+git://github.com/adafruit/Adafruit_Python_DHT.git